# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "G-STK Back-End OpenApi  Documentation",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "description": "Login a user with username and password",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "description": "Logout a user",
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "description": "Return logged in user information",
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/auth/password/email": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "description": "Send reset link Response",
                "operationId": "App\\Http\\Controllers\\Api\\Auth\\ForgotPasswordController::sendResetLinkResponse",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/auth/password/reset": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "description": "Send reset Response",
                "operationId": "App\\Http\\Controllers\\Api\\Auth\\ResetPasswordController::sendResetResponse",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AuthResponse": {
                "title": "AuthResponse",
                "description": "AuthResponse",
                "properties": {
                    "access_token": {
                        "title": "access_token",
                        "type": "string"
                    },
                    "token_type": {
                        "title": "token_type",
                        "type": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "AuthResponse"
                }
            },
...
        }
    },
    "security": [
        []
    ]
}